feat: 당첨 결과를 계산하는 기능 추가
LottoWinCalculator 추가 :
- 로또의 당첨번호를 계산하는 클래스입니다.

WinningType 추가 :
- 당첨 종류를 구분하기 위한 Enum입니다.

+++ src/test/java/lotto/UserInputParserTest.java
@@ -20,7 +20,7 @@ class UserInputParserTest {
                 "'32000', 32000",
                 "'529132000', 529132000"
             })
-    void 구매금액_파서_기능_테스트(String rawInput, long expected) {
+    void 구매금액_파서_기능_테스트(String rawInput, int expected) {
         assertEquals(expected, UserInputParser.getParsedInput(rawInput, UserInputType.PURCHASE_COST));
     }
